+++ app/Http/Filters/PaymentReconcileFilter.php
@@ -4,7 +4,7 @@ namespace App\Http\Filters;
 
 use Illuminate\Database\Eloquent\Builder;
 
-class PaymentFilter extends BaseFilter
+class PaymentReconcileFilter extends BaseFilter
 {
 
 }
