+++ app/PaymentType.php
@@ -11,8 +11,8 @@ class PaymentType extends Model
     /**
      * @return \Illuminate\Database\Eloquent\Relations\HasMany
      */
-    public function paymentLists()
+    public function payments()
     {
-        return $this->hasMany(PaymentList::class);
+        return $this->hasMany(Payment::class);
     }
 }
